+++ pkg/webhook/mutating/webhook.go
@@ -1,17 +1,17 @@
-package useridentity
+package mutating
 
 import (
 	"encoding/json"
 	"net/http"
 
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-
 	"github.com/mattbaird/jsonpatch"
-	"github.com/stolostron/multicloud-operators-foundation/pkg/utils"
-	serve "github.com/stolostron/multicloud-operators-foundation/pkg/webhook/serve"
 	v1 "k8s.io/api/admission/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	rbaclisters "k8s.io/client-go/listers/rbac/v1"
 	"k8s.io/klog/v2"
+
+	"github.com/stolostron/multicloud-operators-foundation/pkg/utils"
+	"github.com/stolostron/multicloud-operators-foundation/pkg/webhook/serve"
 )
 
 type AdmissionHandler struct {
