+++ pkg/webhook/validating/webhook_test.go
@@ -1,18 +1,16 @@
-package clusterset
+package validating
 
 import (
 	"reflect"
 	"testing"
 
 	hivev1 "github.com/openshift/hive/apis/hive/v1"
-
 	hivefake "github.com/openshift/hive/pkg/client/clientset/versioned/fake"
 	v1 "k8s.io/api/admission/v1"
 	authorizationv1 "k8s.io/api/authorization/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	kubefake "k8s.io/client-go/kubernetes/fake"
-
 	clienttesting "k8s.io/client-go/testing"
 )
